Project 1 - Challenge/requirements.txt
sqlalchemy
python-decouple==3.4
unidecode
requests
pandas
psycopg2